requirements.txt
flask-apscheduler
Flask==2.3.3
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
Flask-Mail==0.9.1
twilio==8.10.0
APScheduler==3.10.4
python-dotenv==1.0.1
gunicorn
